pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openenv-ai-code-review"
version = "0.1.0"
description = "AI Code Review OpenEnv environment"
requires-python = ">=3.10"
dependencies = [
    "openenv-core[core]>=0.2.2",
    "fastapi==0.115.12",
    "uvicorn>=0.35,<1",
    "openai>=2.7.2,<3",
    "pydantic==2.12.5",
    "PyYAML==6.0.2",
]

[project.scripts]
server = "server.app:main"

[tool.setuptools]
packages = ["env", "server"]
